refactor: uses resemble the package names
